backend/requirements.txt
astroid
autopep8
certifi
chardet
defusedxml
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
gunicorn
idna
lazy-object-proxy
mccabe
oauthlib
pep8
pycodestyle
pylint
python3-openid
pytz
requests-oauthlib
six
sqlparse
typed-ast
urllib3
wrapt
requests
isort

